+++ nbstripout/_utils.py
@@ -1,11 +1,22 @@
 from collections import defaultdict
 import sys
+from typing import NamedTuple, List, Any
+from nbformat import NotebookNode
 
-__all__ = ["pop_recursive", "strip_output", "strip_zeppelin_output", "MetadataError"]
+__all__ = ["pop_recursive", "strip_output", "strip_zeppelin_output", "StripArgs", "MetadataError"]
 
 
 class MetadataError(Exception):
     pass
+
+class StripArgs(NamedTuple):
+    keep_output: bool
+    keep_count: bool
+    extra_keys: List[str] = []
+    drop_empty_cells: bool = False
+    drop_tagged_cells: List[str] = []
+    strip_init_cells: bool = False
+    max_size:int = 0
 
 
 def pop_recursive(d, key, default=None):
@@ -29,7 +40,7 @@ def pop_recursive(d, key, default=None):
     return default
 
 
-def _cells(nb, conditionals):
+def _cells(nb: NotebookNode, conditionals):
     """Remove cells not satisfying any conditional in conditionals and yield all other cells."""
     if nb.nbformat < 4:
         for ws in nb.worksheets:
@@ -44,7 +55,7 @@ def _cells(nb, conditionals):
             yield cell
 
 
-def get_size(item):
+def get_size(item: Any):
     """ Recursively sums length of all strings in `item` """
     if isinstance(item, str):
         return len(item)
@@ -56,7 +67,7 @@ def get_size(item):
         return len(str(item))
 
 
-def determine_keep_output(cell, default, strip_init_cells=False):
+def determine_keep_output(cell: NotebookNode, default, strip_init_cells=False):
     """Given a cell, determine whether output should be kept
 
     Based on whether the metadata has "init_cell": true,
@@ -82,20 +93,19 @@ def determine_keep_output(cell, default, strip_init_cells=False):
     return default
 
 
-def _zeppelin_cells(nb):
+def _zeppelin_cells(nb: NotebookNode):
     for pg in nb['paragraphs']:
         yield pg
 
 
-def strip_zeppelin_output(nb):
+def strip_zeppelin_output(nb: NotebookNode):
     for cell in _zeppelin_cells(nb):
         if 'results' in cell:
             cell['results'] = {}
     return nb
 
 
-def strip_output(nb, keep_output, keep_count, extra_keys=[], drop_empty_cells=False, drop_tagged_cells=[],
-                 strip_init_cells=False, max_size=0):
+def strip_output(nb: NotebookNode, args: StripArgs):
     """
     Strip the outputs, execution count/prompt number and miscellaneous
     metadata from a notebook object, unless specified to keep either the outputs
@@ -103,11 +113,12 @@ def strip_output(nb, keep_output, keep_count, extra_keys=[], drop_empty_cells=Fa
 
     `extra_keys` could be 'metadata.foo cell.metadata.bar metadata.baz'
     """
-    if keep_output is None and 'keep_output' in nb.metadata:
+    keep_output = args.keep_output
+    if args.keep_output is None and 'keep_output' in nb.metadata:
         keep_output = bool(nb.metadata['keep_output'])
 
     keys = defaultdict(list)
-    for key in extra_keys:
+    for key in args.extra_keys:
         if '.' not in key or key.split('.')[0] not in ['cell', 'metadata']:
             sys.stderr.write(f'Ignoring invalid extra key `{key}`\n')
         else:
@@ -119,13 +130,13 @@ def strip_output(nb, keep_output, keep_count, extra_keys=[], drop_empty_cells=Fa
 
     conditionals = []
     # Keep cells if they have any `source` line that contains non-whitespace
-    if drop_empty_cells:
+    if args.drop_empty_cells:
         conditionals.append(lambda c: any(line.strip() for line in c.get('source', [])))
-    for tag_to_drop in drop_tagged_cells:
+    for tag_to_drop in args.drop_tagged_cells:
         conditionals.append(lambda c: tag_to_drop not in c.get("metadata", {}).get("tags", []))
 
     for cell in _cells(nb, conditionals):
-        keep_output_this_cell = determine_keep_output(cell, keep_output, strip_init_cells)
+        keep_output_this_cell = determine_keep_output(cell, keep_output, args.strip_init_cells)
 
         # Remove the outputs, unless directed otherwise
         if 'outputs' in cell:
@@ -133,10 +144,10 @@ def strip_output(nb, keep_output, keep_count, extra_keys=[], drop_empty_cells=Fa
             # Default behavior (max_size == 0) strips all outputs.
             if not keep_output_this_cell:
                 cell['outputs'] = [output for output in cell['outputs']
-                                   if get_size(output) <= max_size]
+                                   if get_size(output) <= args.max_size]
 
             # Strip the counts from the outputs that were kept if not keep_count.
-            if not keep_count:
+            if not args.keep_count:
                 for output in cell['outputs']:
                     if 'execution_count' in output:
                         output['execution_count'] = None
@@ -144,9 +155,9 @@ def strip_output(nb, keep_output, keep_count, extra_keys=[], drop_empty_cells=Fa
             # If keep_output_this_cell and keep_count, do nothing.
 
         # Remove the prompt_number/execution_count, unless directed otherwise
-        if 'prompt_number' in cell and not keep_count:
+        if 'prompt_number' in cell and not args.keep_count:
             cell['prompt_number'] = None
-        if 'execution_count' in cell and not keep_count:
+        if 'execution_count' in cell and not args.keep_count:
             cell['execution_count'] = None
 
         # Always remove some metadata
